iopex/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
asgiref==3.8.1
asttokens==2.4.1
async-timeout==4.0.3
asyncio-mqtt==0.13.0
attrs==23.2.0
build==1.1.1
cachetools==5.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
decorator==5.1.1
Django==4.2.16
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
et-xmlfile==1.1.0
exceptiongroup==1.2.1
executing==2.0.1
feature-flag==1.2.0
frozenlist==1.4.1
google-api-core==2.19.2
google-api-python-client==2.64.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
googleapis-common-protos==1.65.0
httplib2==0.22.0
idna==3.7
importlib-metadata==7.0.1
ipython==8.18.1
jedi==0.19.1
kafka-python==2.0.2
matplotlib-inline==0.1.7
multidict==6.0.5
numpy==1.26.4
openpyxl==3.1.5
packaging==23.2
paho-mqtt==1.6.1
pandas==2.2.2
parso==0.8.4
pexpect==4.9.0
pip-tools==7.4.0
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==3.20.3
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.2
pydantic_core==2.18.3
Pygments==2.18.0
PyJWT==2.9.0
pymssql==2.3.1
pyodbc==5.1.0
pyparsing==3.1.4
pyproject_hooks==1.0.0
python-dateutil==2.9.0.post0
python-dotenv==0.17.1
pytz==2024.1
quartic-tahu==1.0.0
requests==2.32.2
rsa==4.9
six==1.16.0
sqlparse==0.5.1
stack-data==0.6.3
tomli==2.0.1
traitlets==5.14.3
typing_extensions==4.12.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
yarl==1.9.4
zipp==3.17.0
